add uniTest.java and modifiy SHA1CheckSum.java
